[kbd_v2] Make sure the LEDs are set correctly after initialization
If we received an LED update immediately after USB configuration, before
wait_for_usb_init() returned, the LEDs might be left in the wrong state
when wait_for_usb_init() returns.  I haven't seen this problem occur in
practice yet, but this adds an extra LED refresh to ensure that the
state is updated correctly.