[kbd_v2] Update descriptors, and allow disabling the debug interface
This renumbers the keyboard and debug endpoints (to 1 and 2), and also
adds a flag to disable the USB debug interface at build time.

The flag to disble the debug interface was mainly to use while testing
my keyboard with an IOGear (ATEN) KVM switch.  The KVM switch has a
number of issues which cause the debug interface not to work when
plugged into its keyboard port.

The KVM intercepts and proxies all USB traffic to the keyboard.  It
doesn't seem to reliably forward IN packets to the debug interface--I
can get one initial response, but nothing more after that.  It also
incorrectly forwards SET_REPORT packets for the debug interface to the
keyboard interface instead.