[kbd] Perform intelligent ghosting resolution when diodes are present
My v1 keyboard already has diodes on most of the modifier keys, and I
have added diodes on the shift keys for my v2 keyboard.

This updates the code to be aware of the diodes, and to perform more
intelligent ghosting decisions when the ghosting involves keys with
diodes.  When a diode is present we can usually tell if a detected key
press was a real press or just ghosting.