Add UsbController::StateCallback
This adds a StateCallback to the USB controller.  The controller now
monitors suspend and wake events, and notifies the callback of
suspend/wake/configure/unconfigure.

This also refactors the LED control code in kbd_v2 a bit more, to remove
all hardware-specific code from the Controller class.  This does remove
the idle blink pattern while waiting for USB to be configured, but it is
a bit more informative about when the configure event is received and
when suspend/wake happens.  (A suspend interrupt does appear to be
triggered briefly during initial configuration.)