[kbd] switch to a 4MHz build, and change bounce_ms to 0
A 4Mhz build currently results in approximately 5ms per scan iteration.
My cherry MX key switches are rated for a 5ms bounce time, so this is
about right with no additional delay.  Switching from 1MHz to 4MHz
increases power utilization slightly by about 8mA.

In the future it would be nicer to use a counter to ensure that we wait
for the appropriate bounce time between scan iterations.