[kbd] Ensure rectangle is still detected before de-ghosting
Before performing 1-diode ghosting resolution, make sure the entire
rectangle is still detected.  Finding the rectangle takes a non-zero
amount of time, and it is possible that the full rectangle is no longer
pressed by the time ghost1Diode() runs.  If this occurs, ghost1Diode()
is operating under incorrect assumptions and may decide to block the
wrong key.

This changes ghost1Diode() to rescan the rectangle before proceeding.
This reduces the window of time for this race condition to occur.  In
practice I haven't seen it be an issue any more.

Ideally we should do the same for the 2- and 3-diode cases as well.
I haven't done so yet since I don't have any cases to test this with on
my current keyboard.